requirements.txt
flask==1.1.1
google-cloud-ndb==1.1.2

